requirements.txt
chromadb-client==1.0.12

fastapi
langchain
langchain_chroma
langchain_google_genai
langchain_huggingface
haystack-ai
numpy
openai==1.86.0
pandas==2.3.0
protobuf==3.20.3
pydantic==2.11.5
pymongo==4.13.0
python-dotenv==1.1.0
scikit_learn==1.7.0
sentence_transformers==4.1.0
torch==2.7.1
uvicorn==0.34.3
google-generativeai
python-multipart
pydantic[email]
bcrypt